move isLoggingEnabled, enableLogging, saveNotificationToken to global scope
