+++ cli/src/main/kotlin/com/wire/kalium/cli/RegisterAccountApp.kt
@@ -3,12 +3,12 @@ package com.wire.kalium.cli
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.option
 import com.wire.kalium.logger.KaliumLogLevel
-import com.wire.kalium.logger.KaliumLogger
 import com.wire.kalium.logic.CoreLogic
 import com.wire.kalium.logic.configuration.server.ServerConfig
 import com.wire.kalium.logic.feature.register.RegisterParam
 import com.wire.kalium.logic.feature.register.RegisterResult
 import com.wire.kalium.logic.feature.register.RequestActivationCodeResult
+import com.wire.kalium.network.NetworkLogger
 import kotlinx.coroutines.runBlocking
 import java.util.Scanner
 import kotlin.properties.Delegates
@@ -29,7 +29,7 @@ class RegisterAccountApp : CliktCommand() {
     private var code by Delegates.notNull<Int>()
 
     override fun run(): Unit = runBlocking {
-        KaliumLogger.Config(KaliumLogLevel.DEBUG, "")
+        NetworkLogger.setLoggingLevel(KaliumLogLevel.DEBUG)
         when (val requestResult = requestCode()) {
             is RequestActivationCodeResult.Failure.Generic -> {
                 echo(requestResult.failure)
